+++ jeod_helpers/__init__.py
@@ -18,7 +18,7 @@ __all__ = [
     "events",
     "builder",
 ]
-__version__ = "0.1.3"
+__version__ = "0.1.4"
 
 logger = logging.getLogger(__name__)